Moved all deprecated endpoints to a dedicated controller and ensure they are all marked for removal by v0.9.0
